requirements.txt
pandas
selenium
# notebook
flask
Flask-Cors
python-dotenv
openpyxl
gunicorn